修改退款接口

+++ xzshop-service/src/main/java/com/cdzg/xzshop/vo/admin/refund/GoodsInfoVO.java
@@ -30,7 +30,7 @@ public class GoodsInfoVO {
      * 数量
      */
     @ApiModelProperty(value = "数量")
-    private BigDecimal goodsCount;
+    private Integer goodsCount;
     /**
      * 价格
      */
